// rivetkit-typescript/packages/rivetkit/src/common/bare/generated/client-protocol/v3.ts
export type u64 = bigint;
export type uint = bigint;

export type Cbor = ArrayBuffer;

export function readCbor(bc: bare.ByteCursor): Cbor {
	return bare.readData(bc);
}

export function writeCbor(bc: bare.ByteCursor, x: Cbor): void {
	bare.writeData(bc, x);
}

export type Init = {
	readonly actorId: string;
	readonly connectionId: string;
};
// ...
export type Error = {
	readonly group: string;
	readonly code: string;
	readonly message: string;
	readonly metadata: Cbor | null;
	readonly actionId: uint | null;
};
// ...
export type ActionResponse = {
	readonly id: uint;
	readonly output: Cbor;
};

export function readActionResponse(bc: bare.ByteCursor): ActionResponse {
	return {
		id: bare.readUint(bc),
		output: readCbor(bc),
	};
}

export function writeActionResponse(
	bc: bare.ByteCursor,
	x: ActionResponse,
): void {
	bare.writeUint(bc, x.id);
	writeCbor(bc, x.output);
}

export type Event = {
	readonly name: string;
	readonly args: Cbor;
};

export function readEvent(bc: bare.ByteCursor): Event {
	return {
		name: bare.readString(bc),
		args: readCbor(bc),
	};
}

export function writeEvent(bc: bare.ByteCursor, x: Event): void {
	bare.writeString(bc, x.name);
	writeCbor(bc, x.args);
}

export type ToClientBody =
	| { readonly tag: "Init"; readonly val: Init }
	| { readonly tag: "Error"; readonly val: Error }
	| { readonly tag: "ActionResponse"; readonly val: ActionResponse }
	| { readonly tag: "Event"; readonly val: Event };
// ...
export type ToClient = {
	readonly body: ToClientBody;
};
// ...
export type ActionRequest = {
	readonly id: uint;
	readonly name: string;
	readonly args: Cbor;
};

export function readActionRequest(bc: bare.ByteCursor): ActionRequest {
	return {
		id: bare.readUint(bc),
		name: bare.readString(bc),
		args: readCbor(bc),
	};
}

export function writeActionRequest(
	bc: bare.ByteCursor,
	x: ActionRequest,
): void {
	bare.writeUint(bc, x.id);
	bare.writeString(bc, x.name);
	writeCbor(bc, x.args);
}

export type SubscriptionRequest = {
	readonly eventName: string;
	readonly subscribe: boolean;
};

export function readSubscriptionRequest(
	bc: bare.ByteCursor,
): SubscriptionRequest {
	return {
		eventName: bare.readString(bc),
		subscribe: bare.readBool(bc),
	};
}

export function writeSubscriptionRequest(
	bc: bare.ByteCursor,
	x: SubscriptionRequest,
): void {
	bare.writeString(bc, x.eventName);
	bare.writeBool(bc, x.subscribe);
}

export type ToServerBody =
	| { readonly tag: "ActionRequest"; readonly val: ActionRequest }
	| {
			readonly tag: "SubscriptionRequest";
			readonly val: SubscriptionRequest;
	  };
// ...
export type ToServer = {
	readonly body: ToServerBody;
};
// ...
export type HttpActionRequest = {
	readonly args: Cbor;
};
// ...
export type HttpActionResponse = {
	readonly output: Cbor;
};
// ...
export type HttpQueueSendRequest = {
	readonly body: Cbor;
	readonly name: string | null;
	readonly wait: boolean | null;
	readonly timeout: u64 | null;
};
// ...
export type HttpQueueSendResponse = {
	readonly status: string;
	readonly response: Cbor | null;
};
// ...
export type HttpResponseError = {
	readonly group: string;
	readonly code: string;
	readonly message: string;
	readonly metadata: Cbor | null;
};
// ...
export type HttpResolveResponse = {
	readonly actorId: string;
};